create superclasses and use attribute reader create superclass for both partner and and applicant for both bank_statements controllers and bank_statements forms
